Added edited files
Added collision by storing all blocks exposed to air in a given chunk and checking if they collide with the player. I improved the chunk data structure but changed it to an unordered map for better performance, and since it does not need to be in order since I'm accessing chunks using the player's given position. I also created a struct that stores the vertex array for the terrain and an ordered map for the exposed blocks for collision, so the chunk data structure can carry two bits of essential information. I then added an easy way of switching to Noclip and Flight modes for testing. The collision does need to be improved, though, since you can clip through the map depending on the angle.